+++ brokerapi/brokers/gcp_service_broker.go
@@ -111,12 +111,14 @@ func New(cfg *config.BrokerConfig, Logger lager.Logger) (*GCPAsyncServiceBroker,
 			ProjectId:      cfg.ProjectId,
 			Logger:         self.Logger,
 			AccountManager: sqlManager,
+			SaAccountManager: saManager,
 		},
 		models.CloudsqlPostgresName: &cloudsql.CloudSQLBroker{
 			HttpConfig:     cfg.HttpConfig,
 			ProjectId:      cfg.ProjectId,
 			Logger:         self.Logger,
 			AccountManager: sqlManager,
+			SaAccountManager: saManager,
 		},
 		models.BigtableName: &bigtable.BigTableBroker{
 			BrokerBase: bb,
@@ -316,26 +318,16 @@ func (gcpBroker *GCPServiceBroker) Bind(instanceID string, bindingID string, det
 			err)
 	}
 
-	var creds map[string]string
-	if err := json.Unmarshal([]byte(newCreds.OtherDetails), &creds); err != nil {
-		return models.Binding{}, err
-	}
-
-	// copy provision.otherDetails to creds.
+	// get existing service instance details
 	var instanceRecord models.ServiceInstanceDetails
 	if err = db_service.DbConnection.Where("id = ?", instanceID).First(&instanceRecord).Error; err != nil {
 		return models.Binding{}, fmt.Errorf("Error retrieving service instance details: %s", err)
 	}
 
-	var instanceDetails map[string]string
-	// if the instance has access details saved
-	if instanceRecord.OtherDetails != "" {
-		if err := json.Unmarshal([]byte(instanceRecord.OtherDetails), &instanceDetails); err != nil {
-			return models.Binding{}, err
-		}
+	updatedCreds, err := gcpBroker.ServiceBrokerMap[serviceId].BuildInstanceCredentials(newCreds, instanceRecord)
+	if err != nil {
+		return models.Binding{}, err
 	}
-
-	updatedCreds := gcpBroker.ServiceBrokerMap[serviceId].BuildInstanceCredentials(creds, instanceDetails)
 
 	return models.Binding{
 		Credentials:     updatedCreds,