+++ examples/go/sb-test/api.go
@@ -26,7 +26,6 @@ import (
 	storage "google.golang.org/api/storage/v1"
 	instancepb "google.golang.org/genproto/googleapis/spanner/admin/instance/v1"
 
-	"encoding/pem"
 	_ "github.com/lib/pq"
 )
 
@@ -302,43 +301,15 @@ func testCloudPostgreSQL(w http.ResponseWriter, req *http.Request) {
 	}
 
 	credsInterface := asMap["google-cloudsql-postgres"][0]["credentials"]
-	caCert := credsInterface.(map[string]interface{})["CaCert"].(string)
-	clientCertStr := credsInterface.(map[string]interface{})["ClientCert"].(string)
-	clientKeyStr := credsInterface.(map[string]interface{})["ClientKey"].(string)
+	connStr := credsInterface.(map[string]interface{})["uri"].(string)
 
-	certpem, err := os.Create("/tmp/clientcert.pem")
-	defer certpem.Close()
-	pem.Encode(certpem, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: []byte(clientCertStr),
-	})
-	keypem, err := os.Create("/tmp/clientkey.pem")
-	defer keypem.Close()
-	pem.Encode(keypem, &pem.Block{
-		Type:  "RSA PRIVATE KEY",
-		Bytes: []byte(caCert),
-	})
-	capem, err := os.Create("/tmp/cacert.pem")
-	defer capem.Close()
-	pem.Encode(capem, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: []byte(caCert),
-	})
-
-	dbHost := credsInterface.(map[string]interface{})["host"].(string)
-	dbUsername := credsInterface.(map[string]interface{})["Username"].(string)
-	dbPassword := credsInterface.(map[string]interface{})["Password"].(string)
-	databaseName := credsInterface.(map[string]interface{})["database_name"].(string)
-	dbPort := "3306"
-
-	connStr := fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=require&sslcert=%s&sslkey=%s&sslrootcert=%s",
-		dbUsername, dbPassword, dbHost, dbPort, databaseName, clientCertStr, clientKeyStr, caCert)
-	_, err = sql.Open("postgres", connStr)
+	_, err := sql.Open("postgres", connStr)
 	if err != nil {
 		respond(w, http.StatusInternalServerError, err.Error())
 		return
 	}
 
+	databaseName := credsInterface.(map[string]interface{})["database_name"].(string)
 	respond(w, http.StatusOK, fmt.Sprintf("I connected to the %s database, yay!", databaseName))
 }