+++ CHANGELOG.md
@@ -128,7 +128,23 @@ any extra access details
 - added PostgreSQL support to CloudSQL (and migrated existing plans)
 - added Datastore support
 
-## [4.0.0] - 2017-07-XX
+
+## [3.5.2] - 2017-10-17
+ 
+- fixed Postgres connection uri
+- added wait for ssl certs to finish being created in sql account manager
+
+## [3.6.0] - 2017-01-03
+ 
+- changed default authorized networks for cloudsql instances from `0.0.0.0/0` to none
+- added optional parameter to cloudsql provision operation to specify authorized networks
+- added service account and key provisioning on cloudsql bind operations
+- updated postgres `uri` field to include ssl certificates
+- added optional parameter to cloudsql bind operation to pass back jdbc formatted `uri` field
+- removed waiting for ssl certs to finish being created in sql account manager
+- changed default number of spanner nodes to 1 in `tile.yml`
+
+## [Pending] - 2018-XX-XX
 
 - added command `get_plan_info` to dump plan information to the console. Advised to run before updating.
 - plan ids are now required and will not be generated if not supplied.
@@ -137,4 +153,3 @@ any extra access details
 - modified the formatting of custom plans to be consistent with that of preconfigured plans
 - modified structure of all catalog-related environment variables - `plans` is now a sub-field of the Service object, 
 and Service objects are defined individually by setting env variables like `GOOGLE_<SERVICE_NAME>`
-